Time: 3 ms (37.57%), Space: 6.1 MB (7.30%) - LeetHub

+++ 670-maximum-swap/670-maximum-swap.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int maximumSwap(long long num)
+    int maximumSwap(int num)
     {
         string numString = to_string(num);
         int n = numString.length();
